+++ src/interface/state/config.rs
@@ -1,9 +1,9 @@
-use crate::{Error, state::StateError};
+use crate::Error;
 use candid::Principal;
 
 // get_controllers
 pub fn get_controllers() -> Result<Vec<Principal>, Error> {
-    let config = crate::state::config::get_config().map_err(StateError::ConfigError)?;
+    let config = crate::config::get_config()?;
 
     Ok(config.controllers.to_vec())
 }