+++ src/macros/memory.rs
@@ -15,6 +15,16 @@ macro_rules! icu_memory_manager {
                 $crate::ic::structures::DefaultMemoryImpl::default(),
             ));
 
+            ///
+            /// MEMORY_COUNTER
+            ///
+            pub static MEMORY_COUNTER: ::std::cell::RefCell<$crate::memory::MemoryCounter> =
+                ::std::cell::RefCell::new(<$crate::memory::MemoryCounter>::init(
+                    MEMORY_MANAGER.with_borrow(|this| {
+                        this.get($crate::ic::structures::memory::MemoryId::new(0))
+                    }
+                ),
+            ));
         }
     };
 }
@@ -22,14 +32,10 @@ macro_rules! icu_memory_manager {
 // // icu_memory_add
 #[macro_export]
 macro_rules! icu_memory_add {
-    ($name:ident, $state:ty, $memory_id:expr) => {
-        const _: () = {
-            if !($memory_id >= 10 && $memory_id <= 19) {
-                panic!("state_init! macro error: $memory_id must be between 10 and 19");
-            }
-        };
-
+    ($name:ident, $state:ty) => {
         thread_local! {
+            let next_memory_id = MEMORY_COUNTER.with_borrow(|this| this.next_memory_id());
+
             pub static $name: ::std::cell::RefCell<$state> = ::std::cell::RefCell::new(<$state>::init(
                 MEMORY_MANAGER.with_borrow(|this| {
                     this.get(::icu::ic::structures::memory::MemoryId::new($memory_id))